+++ src/main/java/com/nexign/dao/ProductRepository.java
@@ -14,7 +14,7 @@ public interface ProductRepository extends JpaRepository<Product, Integer> {
 
     List<Product> findAll();
 
-    Optional<Product> findById(Integer id);
+    Optional<Product> findById(Long id);
 
     Optional<Product> findByProductNameAndProducer(@Param("name") String nameProduct, @Param("producer") String producer);
